# openapi.yaml
openapi: 3.1.0
info:
  title: Chatrooms API
  description: The API for Chatrooms.
  version: 1.0.0
servers:
  - url: 'https://api.chatrooms.udupa.dev'
paths:
...
  /invites/{inviteCode}:
    get:
      tags:
        - Room Invites
      summary: Get information for a specified invite.
      parameters:
        - name: inviteCode
          in: path
          description: The code of the invite to get the information for.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "404":
          description: Invite with given code not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      security:
        - cookieAuth: [ ]
      tags:
        - Room Invites
      summary: Accept an invite with a given code.
      description: Accept an invite with a given code as the currently logged in user.
      parameters:
        - name: inviteCode
          in: path
          description: The code of the invite to accept.
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Successfully accepted invite.
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Invite with given code not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Invite has expired, is used or user is already in room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      security:
        - cookieAuth: [ ]
      tags:
        - Room Invites
      summary: Delete an invite.
      description: Delete a room invite with the specified code.
      parameters:
        - name: inviteCode
          in: path
          description: The code of the invite to delete.
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Successfully deleted invite.
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to delete this invite.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Invite with given code not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rooms/{roomId}/messages:
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - Messages
      summary: Get the messages in a given room.
      description: Get the messages in a given room, from newest to oldest.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to get messages from. Defaults to last 25 messages.
          required: true
          schema:
            type: string
        - name: before
          in: query
          description: Get messages before this ID.
          required: false
          schema:
            type: string
        - name: after
          in: query
          description: Get messages after this ID.
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: The maximum number of messages to return (1-100).
          required: false
          schema:
            type: integer
            format: i32
      responses:
        "200":
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to view messages in this channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      security:
        - cookieAuth: [ ]
      tags:
        - Messages
      summary: Send a new message in a room.
      description: Send a new message in a given room, as the currently logged in user.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to send a message in.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreate"
      responses:
        "200":
          description: Successfully sent message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to send messages in this channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid message content provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rooms/{roomId}/messages/{messageId}:
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - Messages
      summary: Fetch a message from a room.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to fetch the message from.
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          description: The ID of the message to fetch.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully fetched message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to view messages in this channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room or message with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      security:
        - cookieAuth: [ ]
      tags:
        - Messages
      summary: Edit a message in a room.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to edit the message in.
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          description: The ID of the message to edit.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessagePatch"
      responses:
        "204":
          description: Successfully edited message.
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to edit this message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room or message with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      security:
        - cookieAuth: [ ]
      tags:
        - Messages
      summary: Delete a message in a room.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to delete the message in.
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          description: The ID of the message to delete.
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Successfully deleted message.
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to delete this message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room or message with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: betterauth.session_token
  schemas:
...
    NullableUser:
      type: [ object, "null" ]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [ string, "null" ]
      required:
        - id
        - username
        - displayName
        - avatarUrl
      example:
        id: "12345676543245"
        username: horizell
        displayName: Horizell
...
    PartialRoom:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: [ string, "null" ]
        name:
          type: string
        description:
          type: [ string, "null" ]
      required:
        - id
        - creatorId
        - name
        - description
      example:
        id: "12384829543245"
        creatorId: "34734829543245"
        name: General
        description: A general place for conversation or smth :)
...
    Invite:
      type: object
      properties:
        code:
          type: string
        room:
          $ref: "#/components/schemas/PartialRoom"
        creator:
          $ref: "#/components/schemas/NullableUser"
        uses:
          type: integer
          format: int32
        maxUses:
          type: [ integer, "null" ]
          format: int32
        expires:
          type: [ string, "null" ]
          format: date-time
      required:
        - id
        - code
        - room
        - creator
        - uses
        - maxUses
        - expires
...
    MessageCreate:
      type: object
      properties:
        content:
          type: string
      required:
        - content
      example:
        content: "Woah it's a message on Chatrooms! That's crazy"
    MessagePatch:
      type: object
      properties:
        content:
          type: string
      required:
        - content
      example:
        content: "Woah it's a message on Chatrooms!"
    Message:
      type: object
      properties:
        id:
          type: string
        author:
          $ref: "#/components/schemas/NullableUser"
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        editedTimestamp:
          type: [ string, "null" ]
          format: date-time
      required:
        - id
        - author
        - content
        - timestamp
        - editedTimestamp
    Error:
      type: object
      properties:
        code:
          type: integer
          format: i32
        message:
          type: string
      example:
        code: 404
        message: User not found.
...